Update OpenAmp Arduino implementation to MP1 Cube version 1.5.0
Merge udates from:
STM32CubeMP1\Projects\STM32MP157C-DK2\Applications\OpenAMP\OpenAMP_TTY_echo\

+++ cores/arduino/stm32/OpenAMP/openamp_conf.h
@@ -54,9 +54,9 @@ extern int __OPENAMP_region_end__[];    /* defined by linker script */
 #define SHM_SIZE                (size_t)((void *)__OPENAMP_region_end__ - (void *) __OPENAMP_region_start__)
 
 #if defined STM32MP1xx
-#define VRING_RX_ADDRESS        -1        /* allocated by Master processor: CA7 */
-#define VRING_TX_ADDRESS        -1        /* allocated by Master processor: CA7 */
-#define VRING_BUFF_ADDRESS      -1        /* allocated by Master processor: CA7 */
+#define VRING_RX_ADDRESS     ((unsigned int)-1)  /* allocated by Master processor: CA7 */
+#define VRING_TX_ADDRESS     ((unsigned int)-1)  /* allocated by Master processor: CA7 */
+#define VRING_BUFF_ADDRESS   ((unsigned int)-1)  /* allocated by Master processor: CA7 */
 #define VRING_ALIGNMENT         16        /* fixed to match with linux constraint */
 #ifndef VRING_NUM_BUFFS
 #define VRING_NUM_BUFFS         16        /* number of rpmsg buffer */
